// src/app/enums/modifier-type.enum.ts
export enum ModifierType {
    AbilityScore_Change = "Change AbilityScore",
    Defense_Add = "Add Defense",
    HitPointMaximum_Change = "Change Hit Point Maximum - Variable",
    Influence_Condition_Add = "Add Condition Influence",
    Influence_Custom_Add = "Add Custom Influence",
    Influence_DamageType_Add = "Add Damage Type Influence",
    Proficiency_Add = "Add Proficiency",
    Sense_Add = "Add Sense",
    Skill_Proficient = "Add Skill Proficiency",
    Speed_Add = "Add Speed",
    Spell_Add = "Add Spell"
}